# server/swagger.yaml
swagger: '2.0'
info:
  description: Mergin API to synchronize your GIS data.
  version: '0.6'
  title: Mergin API
basePath: /v1
...
schemes:
  - https
  - http
definitions:
...
  UserDetail:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      username:
        type: string
        example: 'johnny'
      disk_usage:
        type: integer
        format: int64
        example: 64052
      storage_limit:
        type: integer
        format: int64
        example: 104857600
  LoginResponse:
    allOf:
      - $ref: '#/definitions/UserDetail'
      - type: object
        properties:
          token:
            type: string
            example: '.eJyrVkrNTczMUbJSyk0sSk7NccgpLSlKTM7PKy7NKcnMS9dLztcrzVbSUSotTi2Kz0xRsjKCsPMSc1PhupRqAbKNGXw.XMvndw.XeJ1F7ch2zQvNVEz_zvIPgj4iWY'
          exipre:
            type: string
            format: date-time
            example: '2019-05-04T14:21:56.695035Z'
...
paths:
...
  /user/{username}:
    get:
      deprecated: true
      tags:
        - user
      summary: User profile info.
      description: Returns user profile info.
      operationId: get_user
      parameters:
        - name: username
          in: path
          description: Username.
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Success.
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.user_controller
  /auth/login:
    post:
      deprecated: true
      tags:
        - login
      summary: Login user.
      description: Returns session token, expiration time and user profile info.
      operationId: login
      parameters:
        - name: credentials
          in: body
          description: User credentials.
          required: true
          schema:
            type: object
            required:
              - login
              - password
            properties:
              login:
                type: string
                example: johnny
              password:
                type: string
                example: pass123
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Success.
          schema:
            $ref: '#/definitions/LoginResponse'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
        '415':
          $ref: '#/responses/UnsupportedMediaType'
      x-swagger-router-controller: src.controllers.user_controller
...
responses:
  UnauthorizedError:
    description: Authentication information is missing or invalid.
    headers:
      WWW_Authenticate:
        type: string
  NotFoundResp:
    description: Project not found.
  BadStatusResp:
    description: Invalid request.
...
  UnsupportedMediaType:
    description: Payload format is in an unsupported format.